Change copper weapon cost to reflect size/weight
Copper knife has been brought down to close to it's weight, rounded up.
Copper spear and pike are down to reasonable levels, though i can't find any info on old copper spearheads. I based the cost off of the weight of old bronze spearheads, which are anywhere from the 260-660g range. for 2 copper scrap the heads would be about 860g, which is pretty weighty for a spearhead, but miles better than using 15.